soft number
A soft number is a number that is composed of only even digits (0, 2, 4, 6, 8). In a given range (the beginning m and the end of the range n are integers whose value is entered from the keyboard), to find and print the smallest "soft number". If no such number exists, print NE.