Palindrome divisible by the sum of digits
Make a web page that for a four-digit number entered from a keyboard (via prompt window) will check if it is a palindrome and if it is divisible by the sum of its digits. If the entered number is not four digits, a message should be printed in the alert window that it is not possible check the entered number, otherwise in a message also in the alert window it is written whether it is palindrome and whether it is divisible.