BMI
Create a web page in which there will be 2 input fields (with an explanation of what needs to be entered in these fields) and a Calculate button, for entering weight (in kilograms) and height (in
centimeters) per person. When clicking the Calculate button, the person's BMI should be calculated and written in the alert window. 
Formula: BMI = kg/m2(weight in kilograms divided by height in
square meters.)