Dog years
Write a program to calculate the age of a dog in human years. The dog’s age is read from standard input. If a negative number is entered for the age, print the following message: “Age must be a positive number”.

Note: For the first two years, one dog year is equal to 10.5 human years. After that, each dog year is equal to 4 human years.