Count up and count down
Write a recursive function count_down(int n) that for a given integer n prints the numbers from n to 0.
Write a recursive function count_up(int n) that for a given integer n prints the numbers from 0 to n. (Use the code of the function count_down(int n) and solve the task by changing the order of the commands).