Prime numbers
Write a program that for a given natural number, calculates the difference between the nearest larger prime number and the number itself. The program should use a recursive function to find the appropriate prime number, which in turn should use a recursive function to check if a given number is a prime number.