Sharpened number
Write a program that, for a five-digit number entered from standard input, will 'sharpen' it. A number is 'sharpened' if its leftmost and rightmost digits are removed. Output the sharpened number, and in the second line, print the sum of the digits that were sharpened.

Explanation: The number 54321 is sharpened to 432. The digits that are sharpened are 5 and 1.