Interesting number
A natural number is "interesting" if its reciprocal is divisible by its number of digits. A reversed number is a number made up of the same digits, but in reverse order (for example, 653 is the reverse of the number 356). A natural number n ( n > 9) is entered from the keyboard. To find and print the largest natural number less than n that is "interesting". If the entered number is not valid, print an appropriate message.