Sum of all numbers hidden in string
Characters are read from standard input until an exclamation point is read. Whole numbers (less than 100) are hidden in the text entered like this. To write a program that will read all the characters and print the sum of all the numbers hidden in the text.

Note: cin in C++ ignores spaces. In order not to ignore them, it is necessary to add noskipws before reading the character (eg cin >> noskipws >> znak;)