Pairs of numbers
Write a program in which one positive integer z is first entered from standard input, and then pairs of integers (a, b) are entered successively. Entering pairs of integers should end when the user enters the pair (0, 0). 

It is necessary to calculate how many times z is equal to the sum of each entered pair of numbers a and b, as well as what percentage of the total number of entered pairs (a, b) gives the sum z (NOTE: the pair (0, 0) is not taken into taken into account when performing the calculations!).