Hexadecimal digits
An unknown number of hexadecimal digits are read from standard input until a dot (.) is entered. Your task is to calculate the decimal sum of the entered hexadecimal digits. If the obtained decade sum is divisible by 16, "Pogodok" is printed. If that same sum, besides being divisible by 16 plus, ends at 16 (its last digits are 1 and 6), "Poln pogodok" is printed, otherwise the sum itself is printed.