Index of a max digit in a number
An unknown number of positive integers (with a maximum of 5 digits) are read from the keyboard until something other than a number is entered. For each number, the position (from right to left) of the first occurrence of the largest digit in the composition of the number should be found (the units digit is located at position 0). Then, to print statistics for the positions of the digits with the maximum value found.