Sum of the digits and the max digit of a number
An unknown number of positive integers are read from the keyboard until something other than a number is entered. For each number, the sum of the digit with the maximum value of the previously entered number and the sum of the digits of the current number should be calculated (for the first entered number, only the sum of its digits is calculated). For each entered number the result should be printed in the following format:

[number]: [sum of the digits of the number + the maximum digit of the previous number]