Adjacent vowels
Strings are entered from standard input. To count and print all final occurrences of adjacent vowels in sentences. Ignore the occurrence of upper and lower case letters. Print the found vowel pairs to the screen, each on a new line in lowercase. Then the number of occurrences of vowel pairs is printed in a new line. Reading ends when the # character is read