Max digit in an array of numbers
From standard input is given multiple rows of numbers such that each row starts with a single integer (N>=1) indicating how many numbers follow it in the same row. To write a program that will print the number with the largest most significant digit on SI for each row. Reading of numbers ends when the number 0 is read.