The sum of a given index
Write a sum_pos function that receives the array, the number ind (which represents an index in the array), and the number of valid elements in the array. The function should return the sum of all elements starting from the given index ind and ending at the end of the array. If the index ind is greater than N the function should return 0.