Quadrants
Find the sums of each of the quadrants and print them to standard output. The sum is printed for the first quadrant, then for the second, then for the third and finally for the fourth. If a quadrant cannot be created, then 0 should be printed for the sum of that quadrant.