Sport ticket
A sportsbook ticket is read from a standard input. In the first line, the payment amount is written (whole number).
Then in each line read from standard input (until the # character is read) one type is written in the following format:
ab12 1 1.25
The first number is the type code (a string of characters no longer than 9 characters), the second number is the type (can be 1, 0 or 2) while the third number is the coefficient (a real number). Your task is to print the type with the highest odds as well as the possible win on the ticket. If there are several types with the same maximum coefficient, print the first one.