Consecutive 1s in a matrix
Write a program that loads a matrix with dimensions MxN (max. 100x100). At the beginning, the dimensions of the matrix are entered, and then the elements of the matrix, which are only the values ​​1 and 0. The program should count and print on SI how many of the rows and columns have at least 3 consecutive elements with the value 1.