Element in a matrix
The dimensions of a matrix (m, n <= 100) and then the elements of the matrix are entered from the keyboard. Generate an array (of at most m) such that each element of the array is obtained by finding the element in each row of the matrix that is farthest from the arithmetic mean within that row. If there are multiple elements that are farthest from the arithmetic mean, then the first one is taken into account. The order of writing the elements in the array corresponds to the order of the queues.