Characters sorted in ascending order
Strings of characters are read from standard input until the # character is read. To write a program in which, for each line read, the number of digits in that line will be printed, the sign :, then the digits themselves arranged according to the ASCII code in ascending order. Character strings are no longer than 100 characters.