Recursive function for positive integers in an array
Write a recursive function that will find the number of positive numbers from an integer sequence. The function receives as an argument the array, for which the number of positive numbers and the total number of elements that the array has are required. The function is defined with the following prototype:

int NumberPositive(int array[], int n);