Change the length of rows and columns
Write a program that loads a matrix of integers A with N rows and N * 2 columns (N is not greater than 50). The program should transform the loaded matrix into a new matrix B, so that all elements to the right of the Nth column will be transferred under the Nth row, that is, from a matrix with dimension N rows and 2 * N columns, a matrix with 2 * N rows x N columns.