Replace middle element of each row
Transform the matrix so that the middle element in each row is replaced by the difference (in absolute value) of the sum of the elements in the first half of the row and the sum of the elements in the second half of the row. If the matrix has an even number of columns, the value of the middle two elements is changed. The middle element(s) enters the sums (in the case of an odd number of columns, the middle element enters both sums!). To print the changed matrix on the screen.